+++ src/app/docs/platforms/cli/page.tsx
@@ -96,3 +96,5 @@ goose session`}</code>
 
 
 
+
+